+++ spec/support/cookbooks/resource_test/libraries/resource/python_package.rb
@@ -13,13 +13,7 @@ class Chef
       provides :python_package
 
       property :package_name, [String, Array], identity: true
-      property :group, [String, Integer]
-      property :install_options, [String, Array]
-      property :list_options, [String, Array]
       property :python, String
-      property :user, [String, Integer]
-      property :version, String
-      property :virtualenv, String
 
       %i[install upgrade remove].each do |act|
         action act do