+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of the snu_python cookboo
 ## 1.0.0 (2018-07-11)
 
 - Delete all custom resources except snu_python
+- Remove awscli from the default Python 2 packages
 
 ## 0.2.1 (2018-08-03)